Trait sc_consensus_babe_rpc::BabeApiServer

pub trait BabeApiServer: Sized + Send + Sync + 'static {
    // Required method
    fn epoch_authorship<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = RpcResult<HashMap<AuthorityId, EpochAuthorship>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;

    // Provided method
    fn into_rpc(self) -> RpcModule<Self> { ... }
}

Server trait implementation for the BabeApi RPC API.

Required Methods

fn epoch_authorship<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = RpcResult<HashMap<AuthorityId, EpochAuthorship>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Returns data about which slots (primary or secondary) can be claimed in the current epoch with the keys in the keystore.

Provided Methods

fn into_rpc(self) -> RpcModule<Self>

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.